# Breaking up ReinterpretArray methods by typcasting type to enable more optimizations