--- a/base/reinterpretarray.jl
+++ b/base/reinterpretarray.jl
@@ -566,16 +566,16 @@ end
 
 @inline @propagate_inbounds function setindex!(a::ReinterpretArray{T,N,S}, v, inds::Vararg{Int, N}) where {T,N,S}
     check_writable(a)
-    @inline _setindex_ra!(a, v, inds[1], tail(inds))
+    @inline _setindex_ra!(TypecastStyle(T,S), a, v, inds[1], tail(inds))
 end
 
 @inline @propagate_inbounds function setindex!(a::ReinterpretArray{T,N,S}, v, i::Int) where {T,N,S}
     check_writable(a)
     if isa(IndexStyle(a), IndexLinear)
-        return _setindex_ra!(a, v, i, ())
+        return _setindex_ra!(TypecastStyle(T,S), a, v, i, ())
     end
     inds = _to_subscript_indices(a, i)
-    @inline _setindex_ra!(a, v, inds[1], tail(inds))
+    @inline _setindex_ra!(TypecastStyle(T,S), a, v, inds[1], tail(inds))
 end
 
 @inline @propagate_inbounds function setindex!(a::ReshapedReinterpretArray{T,N,S}, v, ind::SCartesianIndex2) where {T,N,S}
@@ -590,7 +590,93 @@ end
     return a
 end
 
-@inline @propagate_inbounds function _setindex_ra!(a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToSimilar, a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    if issingletontype(T) # singleton types
+        @boundscheck checkbounds(a, i1, tailinds...)
+        # setindex! is a noop except for the index check
+    else
+        setindex!(a.parent, reinterpret(S, v), i1, tailinds...)
+    end
+    return a
+end
+
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToLarger, a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    ind_start = div((i1-1)*sizeof(T), sizeof(S))
+    # Optimizations that avoid branches
+    # T is bigger than S and contains an integer number of them
+    t = Ref{T}(v)
+    GC.@preserve t begin
+        sptr = Ptr{S}(unsafe_convert(Ref{T}, t))
+        n = sizeof(T) ÷ sizeof(S)
+        for i = 1:n
+            s = unsafe_load(sptr, i)
+            a.parent[ind_start + i, tailinds...] = s
+        end
+    end
+    return a
+end
+
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToSmaller, a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    ind_start, sidx = divrem((i1-1)*sizeof(T), sizeof(S))
+    # S is bigger than T and contains an integer number of them
+    s = Ref{S}(a.parent[ind_start + 1, tailinds...])
+    GC.@preserve s begin
+        tptr = Ptr{T}(unsafe_convert(Ref{S}, s))
+        unsafe_store!(tptr + sidx, v)
+        a.parent[ind_start + 1, tailinds...] = s[]
+    end
+    return a
+end
+
+@inline @propagate_inbounds function _setindex_ra!(::Union{TypecastToLargerInexact,TypecastToSmallerInexact,TypecastToSimilarWithFields,TypecastNone},
+        a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    ind_start, sidx = divrem((i1-1)*sizeof(T), sizeof(S))
+    # Optimizations that avoid branches
+    t = Ref{T}(v)
+    s = Ref{S}()
+    GC.@preserve t s begin
+        tptr = Ptr{UInt8}(unsafe_convert(Ref{T}, t))
+        sptr = Ptr{UInt8}(unsafe_convert(Ref{S}, s))
+        nbytes_copied = 0
+        i = 1
+        # Deal with any partial elements at the start. We'll have to copy in the
+        # element from the original array and overwrite the relevant parts
+        if sidx != 0
+            s[] = a.parent[ind_start + i, tailinds...]
+            nb = min((sizeof(S) - sidx) % UInt, sizeof(T) % UInt)
+            _memcpy!(sptr + sidx, tptr, nb)
+            nbytes_copied += nb
+            a.parent[ind_start + i, tailinds...] = s[]
+            i += 1
+            sidx = 0
+        end
+        # Deal with the main body of elements
+        while nbytes_copied < sizeof(T) && (sizeof(T) - nbytes_copied) > sizeof(S)
+            nb = min(sizeof(S), sizeof(T) - nbytes_copied)
+            _memcpy!(sptr, tptr + nbytes_copied, nb)
+            nbytes_copied += nb
+            a.parent[ind_start + i, tailinds...] = s[]
+            i += 1
+        end
+        # Deal with trailing partial elements
+        if nbytes_copied < sizeof(T)
+            s[] = a.parent[ind_start + i, tailinds...]
+            nb = min(sizeof(S), sizeof(T) - nbytes_copied)
+            _memcpy!(sptr, tptr + nbytes_copied, nb)
+            a.parent[ind_start + i, tailinds...] = s[]
+        end
+    end
+    return a
+end
+#=
+@inline @propagate_inbounds function _setindex_ra!(::TS, a::NonReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {TS<:TypecastStyle,T,N,S,TT}
     v = convert(T, v)::T
     # Make sure to match the scalar reinterpret if that is applicable
     if sizeof(T) == sizeof(S) && (fieldcount(T) + fieldcount(S)) == 0
@@ -662,8 +748,70 @@ end
     end
     return a
 end
+=#
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToSimilar, a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    if issingletontype(T) # singleton types
+        @boundscheck checkbounds(a, i1, tailinds...)
+        # setindex! is a noop except for the index check
+    else
+        setindex!(a.parent, reinterpret(S, v), i1, tailinds...)
+    end
+    return a
+end
 
-@inline @propagate_inbounds function _setindex_ra!(a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToSimilarWithFields, a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    t = Ref{T}(v)
+    GC.@preserve t begin
+        sptr = Ptr{S}(unsafe_convert(Ref{T}, t))
+        s = unsafe_load(sptr)
+        a.parent[i1, tailinds...] = s
+    end
+    return a
+end
+
+@inline @propagate_inbounds function _setindex_ra!(::Union{TypecastToLarger,TypecastToLargerInexact},
+        a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    t = Ref{T}(v)
+    GC.@preserve t begin
+        sptr = Ptr{S}(unsafe_convert(Ref{T}, t))
+        # Extra dimension in the parent array
+        n = sizeof(T) ÷ sizeof(S)
+        if isempty(tailinds) && IndexStyle(a.parent) === IndexLinear()
+            offset = n * (i1 - firstindex(a))
+            for i = 1:n
+                s = unsafe_load(sptr, i)
+                a.parent[i + offset] = s
+            end
+        else
+            for i = 1:n
+                s = unsafe_load(sptr, i)
+                a.parent[i, i1, tailinds...] = s
+            end
+        end
+    end
+    return a
+end
+
+@inline @propagate_inbounds function _setindex_ra!(::TypecastToSmaller, a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {T,N,S,TT}
+    v = convert(T, v)::T
+    @boundscheck checkbounds(a, i1, tailinds...)
+    s = Ref{S}()
+    GC.@preserve s begin
+        tptr = Ptr{T}(unsafe_convert(Ref{S}, s))
+        s[] = a.parent[tailinds...]
+        unsafe_store!(tptr, v, i1)
+        a.parent[tailinds...] = s[]
+    end
+    return a
+end
+
+#= 
+@inline @propagate_inbounds function _setindex_ra!(::TS, a::ReshapedReinterpretArray{T,N,S}, v, i1::Int, tailinds::TT) where {TS<:TypecastStyle,T,N,S,TT}
     v = convert(T, v)::T
     # Make sure to match the scalar reinterpret if that is applicable
     if sizeof(T) == sizeof(S) && (fieldcount(T) + fieldcount(S)) == 0
@@ -711,7 +859,8 @@ end
         end
     end
     return a
-end
+ends
+=#
 
 # Padding
 struct Padding